src/requirements.txt
PyGithub==v2.5.0
packaging==24.2
flask==3.1.0
gevent==24.11.1
